[Refactor] profile 분리
