+++ src/test/java/me/snaptime/user/controller/UserControllerTest.java
@@ -4,11 +4,11 @@ import com.google.gson.Gson;
 import me.snaptime.config.SecurityConfig;
 import me.snaptime.jwt.JwtProvider;
 import me.snaptime.jwt.UserDetailsServiceImpl;
+import me.snaptime.profile.service.impl.ProfileServiceImpl;
 import me.snaptime.user.dto.req.UserReqDto;
 import me.snaptime.user.dto.req.UserUpdateDto;
 import me.snaptime.user.dto.res.UserResDto;
 import me.snaptime.user.service.impl.SignServiceImpl;
-import me.snaptime.user.service.impl.UserProfileServiceImpl;
 import me.snaptime.user.service.impl.UserServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -44,7 +44,7 @@ public class UserControllerTest {
     private SignServiceImpl signService;
 
     @MockBean
-    private UserProfileServiceImpl userProfileService;
+    private ProfileServiceImpl profileService;
 
     @MockBean
     private UserDetailsServiceImpl userDetailsService;
